Fix downstream grenade duping bug

+++ code/modules/projectiles/projectile/projectilegrenades.dm
@@ -12,13 +12,18 @@
 
 /obj/item/projectile/bullet/grenade/Move()	//Makes grenade shells cause their effect when they arrive at their target turf
 	if(get_turf(src) == get_turf(original))
+		if(QDELETED(src)) return
+
 		grenade_effect(get_turf(src))
 		qdel(src)
 	else
 		..()
 
 /obj/item/projectile/bullet/grenade/on_hit(atom/target)	//Allows us to cause different effects for each grenade shell on hit
+	if(QDELETED(src)) return
+
 	grenade_effect(target)
+	qdel(src)
 
 /obj/item/projectile/bullet/grenade/proc/grenade_effect(target)
 	return
